completion of part 4

+++ part4/controllers/blogs.js
@@ -2,8 +2,10 @@ const blogRouter = require('express').Router()
 const mongoose = require('mongoose')
 const Blog = require('../models/blogs')
 const User = require('../models/user')
-const { request, response } = require('../app')
+const info = require('../utils/logger')
+const jwt = require('jsonwebtoken')
 const blogs = require('../models/blogs')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/',async(request,response)=>{
     const blogs = await Blog.find({}).populate('user',{username:1, name: 1})
@@ -25,34 +27,63 @@ blogRouter.get('/:id',async(request,response)=>{
       response.status(500).send({ error: 'Server error' });
   }
 })
-blogRouter.put('/:id',async(request,response)=>{
-  const {title, author,like,url} = request.params
+blogRouter.post('/',userExtractor, async (request, response) => {
+  try {
+    const { title, author, url } = request.body
 
-  blogs.findByIdAndUpdate()
-})
-
-blogRouter.post('/',async(request,response,next)=>{
-        const {title, author,url} = request.body
-        const  userId = request.body.userId
     if (!title || !url) {
-        return response.status(400).json({ error: 'Title and url are required' });
-      }
-    const user = await User.findById(userId)
+      return response.status(400).json({ error: 'Title and url are required' })
+    }
 
-    const blogs = new Blog({title, author, url, user: user.id})
+   const users = request.user
+
+    if (!users.id) {
+      return response.status(401).json({ error: 'Token invalid' })
+    }
+
+
+    const user = await User.findById(users.id)
+    if (!user) {
+      return response.status(404).json({ error: 'User not found' })
+    }
+
+
+    const blogs = new Blog({ title, author, url, user: user.id })
     const savedBlog = await blogs.save()
+
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
+
+
     response.status(201).json(savedBlog)
 
+  } catch (error) {
+    console.error(error)
+    response.status(500).json({ error: 'An error occurred while processing your request' })
+  }
+});
 
-})
+blogRouter.delete('/:id', userExtractor,async (request, response) => {
+  const user = request.user
+  try {
+    const blog = await Blog.findById(request.params.id);
+    
+    if (!blog) {
+      return response.status(404).json({ error: 'Blog not found' });
+    }
+    if ( !blog.user || blog.user.toString() !== user.id.toString()) {
+      return response.status(403).json({ error: 'Unauthorized to delete this blog' });
+    }
 
-blogRouter.delete('/:id',async(request,response)=>{
-   await  Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    await Blog.findByIdAndDelete(request.params.id);
 
-})
+    response.status(204).end();  
+
+  } catch (error) {
+    console.error(error);
+    response.status(500).json({ error: 'An error occurred while deleting the blog' });
+  }
+});
 
 blogRouter.put('/:id',async (request,response)=>{
     const body= request.body
@@ -63,9 +94,8 @@ blogRouter.put('/:id',async (request,response)=>{
         author:body.author,
         url:body.url
     }
-
-   await  Blog.findByIdAndUpdate(id,blogs,{new:true})
-   response.json(updatedBlog)
+   const updatedBlog = await  Blog.findByIdAndUpdate(id,blogs,{new:true})
+   response.status(200).json(updatedBlog)
 })
 
 module.exports = blogRouter
